Pull in new Version of Box2D.
Source: https://github.com/erincatto/Box2D.git
Revision: 6a69ddbbd5

Applied a dos2unix run to switch to unix line endings.

+++ external/Box2D/Collision/Shapes/b2CircleShape.h
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2006-2009 Erin Catto http://www.gphysics.com
+* Copyright (c) 2006-2009 Erin Catto http://www.box2d.org
 *
 * This software is provided 'as-is', without any express or implied
 * warranty.  In no event will the authors be held liable for any damages
@@ -44,7 +44,7 @@ public:
 	void ComputeAABB(b2AABB* aabb, const b2Transform& transform, int32 childIndex) const;
 
 	/// @see b2Shape::ComputeMass
-	void ComputeMass(b2MassData* massData, qreal density) const;
+	void ComputeMass(b2MassData* massData, float32 density) const;
 
 	/// Get the supporting vertex index in the given direction.
 	int32 GetSupport(const b2Vec2& d) const;
